backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
psycopg2-binary==2.9.9
google-cloud-secret-manager==2.20.0
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.52.0
google-cloud-discoveryengine==0.11.11
vertexai==1.52.0
